Data-Service/requirements.txt
asyncua
pyModbusTCP
pysnmp
c104
paho-mqtt
fastapi
uvicorn
python-dotenv
